Add some KeyDown and MouseDown functions to trigger a shot
* Using LogicMgr::MouseDown() seems to be the best way to emulate a
shot.
* Using Board::KeyDown() to emulate an "ENTER" key down event is
problematic, since will only fire the shot if the mouse is on the
game window. Since mods run programmatically, it's not guaranteed
that the mouse will always be on the game screen. Since
LogicMgr::MouseDown() accepts the mouse position as arguments, our
mods can lie about the mouse position to trick the game into
thinking that the mouse is on the game screen.
* Fix wrong (?) function address for LogicMgr::BeginShot().
* Add a helper function to convert from degrees to radians, where
0 degrees is straight down.
* Refactor example mod with a state machine.
* Example mod now shoots left, middle, and right, with each shot
being delayed by 20 seconds.
* There is a bug with the SDK when you load a level, leave that
level at some time, and call an SDK function while in the menu. In
the menus, several objects in the game are undefined. The mod SDK
thinks that the objects still exist, so it tries to call the desired
functions. I have a potential fix in the works, just need to figure
out how to get the pointers to cooperate.