Refine the gun angle usability
* Add functions for setting and getting the gun angle in degrees and
radians.
* Fix bug where the static pointer for the LogicMgr wasn't using the
correct address (forgot to dereference pointer to the pointer).
* Fix radians to degrees conversion function.
* Perform modulo division in the angle conversion functions to keep
everything in the standard range.
* Add a random delay to the first shot of a level when `Example_1()`
example mod is running.
* Fix `Example_1()` when it tries to do the gun angle stuff when the
LogicMgr state is set to an unknown value (unknown when the object
first initializes; led to a race condition).
* Set the angle of the gun at a rate of 60 FPS in `Example_1()` and
crank up the rotation speed. (TODO: replace `Sleep()` with a perf
timer.)