Add initial LogicMgr struct data for Peggle Nights (not yet tested)
* Add the static pointer path for Nights's LogicMgr.
* Set the `None` `PowerupType` enum to 0, so that the rest of the
enum uses a known base number to increment from.
* Add the Electrobolt and the weird AI Player powerup to the
`PowerupType` enum.
* Make `PowerupType` an `enum class` for now, until we get sick of
`enum class`es.
* Add `CharacterId` enum, which might actually be better off in
another class, if appropriate.
* Add `PlayerId` enum. I assume mod devs will just use int literals
for player numbers, but having the enum type makes it easy to
interpret the LogicMgr struct members.
* Replace `struct LogicMgr_` with `struct LogicMgr_Deluxe101_` and
`struct LogicMgr_Nights10_`.
* In the `State` enum, rename `CharacterDialog` to `CharacterSelect`
so it is more intuitive that the state is the character selection
dialog.